task1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -2,14 +2,15 @@
 
 /**
  * array_iterator - iterate the array
- *
  * @array: array
  * @size: size of array
  * @action: pointer to function
+ *
+ * Return: void
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
 	if (array || action)
 		for (i = 0 ; i < size ; i++)
